Move messages to their own modules
