+++ cable/slack_test.go
@@ -1,6 +1,7 @@
 package cable
 
 import (
+	telegram "github.com/go-telegram-bot-api/telegram-bot-api"
 	"github.com/nlopes/slack"
 	. "github.com/stretchr/testify/assert"
 	"testing"
@@ -103,4 +104,61 @@ WAIT:
 
 	second := asSlackJSONMessage(client.sent[1])
 	Equal(t, ":clap: Psss!", second.Text)
+}
+
+func TestSlackMessage_String_KnownUser(t *testing.T) {
+	user := slack.User{ID: slackUserID, RealName: "Will Smith", Name: "freshprince"}
+	msg := createSlackMessage("Sup Jay!", slackUserID, user)
+	Equal(t, "freshprince: Sup Jay!", msg.String())
+}
+
+func TestSlackMessage_String_Stranger(t *testing.T) {
+	user := slack.User{ID: slackUserID, RealName: "Will Smith", Name: "freshprince"}
+	msg := createSlackMessage("Sup Jay!", unknownSlackUSerID, user)
+	Equal(t, "Stranger: Sup Jay!", msg.String())
+}
+
+func TestSlackMessage_ToSlack(t *testing.T) {
+	user := slack.User{ID: slackUserID, RealName: "Will Smith", Name: "freshprince"}
+	msg := createSlackMessage("Sup Jay!", slackUserID, user)
+	_, e := msg.ToSlack()
+	Error(t, e)
+}
+
+func TestSlackMessage_ToTelegram_KnownUser(t *testing.T) {
+	user := slack.User{ID: slackUserID, RealName: "Will Smith", Name: "freshprince"}
+	msg := createSlackMessage("Sup Jay! :boom:", slackUserID, user)
+	telegramChatID := int64(123)
+
+	expected := telegram.MessageConfig{
+		BaseChat: telegram.BaseChat{
+			ChatID:           telegramChatID,
+			ReplyToMessageID: 0,
+		},
+		Text:                  "*Will Smith (freshprince):* Sup Jay! 💥 ",
+		DisableWebPagePreview: false,
+		ParseMode:             "Markdown",
+	}
+
+	actual, _ := msg.ToTelegram(telegramChatID)
+	Equal(t, expected, actual)
+}
+
+func TestSlackMessage_ToTelegram_Stranger(t *testing.T) {
+	user := slack.User{ID: slackUserID, RealName: "Will Smith", Name: "freshprince"}
+	msg := createSlackMessage("Sup Jay! :boom:", "STRGRID", user)
+	telegramChatID := int64(123)
+
+	expected := telegram.MessageConfig{
+		BaseChat: telegram.BaseChat{
+			ChatID:           telegramChatID,
+			ReplyToMessageID: 0,
+		},
+		Text:                  "*Stranger:* Sup Jay! 💥 ",
+		DisableWebPagePreview: false,
+		ParseMode:             "Markdown",
+	}
+
+	actual, _ := msg.ToTelegram(telegramChatID)
+	Equal(t, expected, actual)
 }
